# swagger/swagger_output.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "oneM2M API",
        "description": "oneM2M API 1.0",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default",
            "variables": {}
        }
    ],
    "paths": {
...
        "/api/onem2m/devices/data/{ae}/{cnt}": {
            "post": {
                "tags": [
                    "contentInstance"
                ],
                "description": "contentInstance 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cnt",
                        "in": "path",
                        "description": "CNT resourceName, CNT 리소스의 이름",
                        "example": "Test_CNT_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CIN_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "CIN 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/data/{ae}/{cnt}/{cin}": {
            "get": {
                "tags": [
                    "contentInstance"
                ],
                "description": "contentInstance 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cnt",
                        "in": "path",
                        "description": "CNT resourceName, CNT 리소스의 이름",
                        "example": "Test_CNT_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cin",
                        "in": "path",
                        "description": "CIN resourceName, CIN 리소스의 이름",
                        "example": "4-20211220021030252",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "CIN 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "contentInstance"
                ],
                "description": "contentInstance 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cnt",
                        "in": "path",
                        "description": "CNT resourceName, CNT 리소스의 이름",
                        "example": "Test_CNT_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cin",
                        "in": "path",
                        "description": "CIN resourceName, CIN 리소스의 이름",
                        "example": "4-20211220021030252",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "CIN 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/data/{ae}/{cnt}/latest": {
            "get": {
                "tags": [
                    "contentInstance"
                ],
                "description": "가장 최신 contentInstance 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cnt",
                        "in": "path",
                        "description": "CNT resourceName, CNT 리소스의 이름",
                        "example": "STest_CNT_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "가장 최신 CIN 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/data/{ae}/{cnt}/oldest": {
            "get": {
                "tags": [
                    "contentInstance"
                ],
                "description": "가장 과거 contentInstance 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cnt",
                        "in": "path",
                        "description": "CNT resourceName, CNT 리소스의 이름",
                        "example": "Test_CNT_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "가장 과거 CIN 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/onem2m/devices/group/{ae}/{grp}/data": {
            "get": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 멤버 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "GRP 리소스 멤버 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "description": "Group 리소스 멤버 데이터 제어",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_Group_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grp",
                        "in": "path",
                        "description": "GRP resourceName, GRP 리소스의 이름",
                        "example": "Test_GRP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GRP_Member_Control"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "GRP 리소스 멤버 데이터 제어",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CIN_Create": {
                "type": "object",
                "properties": {
                    "con": {
                        "type": "string",
                        "description": "content, 컨텐츠 (센서 데이터, 위치 정보 등)",
                        "example": "on"
                    },
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": [
                            "lamp_01"
                        ]
                    }
                },
                "required": [
                    "con"
                ]
            },
...
            "GRP_Member_Control": {
                "type": "object",
                "properties": {
                    "con": {
                        "type": "string",
                        "description": "content, 컨텐츠 (센서 데이터, 위치 정보 등)",
                        "example": "off"
                    },
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": "Lamps_01"
                    }
                },
                "required": [
                    "con"
                ]
            },
...
        }
    }
}